requirements.txt
python-dotenv
requests
uvicorn[standard]
fastapi
python-multipart
urllib3==1.26.6